Moved developers_api.py to root directory
